Refactor proof artifacts check into submodules
