+++ xtask/src/tasks/proof_artifacts_check/mod.rs
@@ -3,50 +3,31 @@ use crate::cli::{
     ProofRunArtifactsCheckArgs, ProofRunObservationArgs, ProofRunObservationsSummaryArgs,
 };
 use anyhow::{Context, Result, bail};
-use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use std::collections::{BTreeMap, BTreeSet};
 use std::fs;
 use std::path::{Path, PathBuf};
 use walkdir::WalkDir;
 
-const SUMMARY_SCHEMA: &str = "tokmd.proof_executor_summary.v1";
-const MANIFEST_SCHEMA: &str = "tokmd.proof_executor_manifest.v1";
-const PROOF_RUN_SUMMARY_SCHEMA: &str = "tokmd.proof_run_summary.v1";
-const PROOF_RUN_OBSERVATION_SCHEMA: &str = "tokmd.proof_run_observation.v1";
-const PROOF_RUN_OBSERVATION_COLLECTION_SCHEMA: &str = "tokmd.proof_run_observation_collection.v1";
-const OBSERVATION_SCHEMA: &str = "tokmd.proof_executor_observation.v1";
-const OBSERVATION_COLLECTION_SCHEMA: &str = "tokmd.proof_executor_observation_collection.v1";
-const PROMOTION_READINESS_SCHEMA: &str = "tokmd.proof_executor_promotion_readiness.v1";
-const PROOF_ARTIFACTS_CHECK_SCHEMA: &str = "tokmd.proof_artifacts_check.v1";
-const PROOF_EXECUTION_ARTIFACTS_CHECK_SCHEMA: &str = "tokmd.proof_execution_artifacts_check.v1";
-const PROOF_RUN_ARTIFACTS_CHECK_SCHEMA: &str = "tokmd.proof_run_artifacts_check.v1";
+mod io;
+mod json_fields;
+mod markdown;
+mod receipts;
+mod schemas;
+mod types;
 
-const SHARED_FIELDS: &[&str] = &[
-    "mode",
-    "status",
-    "execution_status",
-    "execution_guard",
-    "family",
-    "required",
-    "profile",
-    "base",
-    "head",
-    "ok",
-    "changed_files",
-    "unknown_files",
-];
-
-const ENTRY_FIELDS: &[&str] = &[
-    "scope",
-    "kind",
-    "required",
-    "command",
-    "artifact_path",
-    "status",
-    "skip_reason",
-    "exit_code",
-];
+use io::{
+    artifact_root_for, read_json, write_observation, write_proof_run_observation, write_text,
+};
+use json_fields::*;
+use markdown::{
+    render_observation_collection_markdown, render_proof_run_observation_collection_markdown,
+};
+#[cfg(test)]
+use receipts::{ProofArtifactsCheckCounts, ProofArtifactsCheckReceipt};
+use receipts::{executor_check_receipt, proof_run_check_receipt, write_check_receipt};
+use schemas::*;
+use types::*;
 
 pub fn run(args: ProofArtifactsCheckArgs) -> Result<()> {
     let outcome = (|| {
@@ -245,528 +226,6 @@ pub fn run_observations_summary(args: ProofExecutionObservationsSummaryArgs) -> 
     Ok(())
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum VerificationMode {
-    NoExecution,
-    Execution,
-}
-
-#[derive(Debug, PartialEq, Eq)]
-struct ProofArtifactsReport {
-    selected: usize,
-    executed: usize,
-    execution_status: String,
-    guard_reason: String,
-}
-
-#[derive(Debug, PartialEq, Eq)]
-struct ProofRunArtifactsReport {
-    executed: usize,
-    guard_reason: String,
-}
-
-#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
-struct ProofArtifactsCheckReceipt {
-    schema: String,
-    ok: bool,
-    verifier: String,
-    inputs: BTreeMap<String, String>,
-    counts: ProofArtifactsCheckCounts,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    execution_status: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    guard_reason: Option<String>,
-    errors: Vec<String>,
-}
-
-#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
-struct ProofArtifactsCheckCounts {
-    selected: usize,
-    executed: usize,
-}
-
-fn executor_check_receipt(
-    schema: &str,
-    verifier: &str,
-    args: &ProofArtifactsCheckArgs,
-    outcome: &std::result::Result<ProofArtifactsReport, anyhow::Error>,
-) -> ProofArtifactsCheckReceipt {
-    let mut inputs = BTreeMap::new();
-    inputs.insert(
-        "executor_summary".to_string(),
-        args.executor_summary.to_string_lossy().to_string(),
-    );
-    inputs.insert(
-        "executor_manifest".to_string(),
-        args.executor_manifest.to_string_lossy().to_string(),
-    );
-
-    match outcome {
-        Ok(report) => ProofArtifactsCheckReceipt {
-            schema: schema.to_string(),
-            ok: true,
-            verifier: verifier.to_string(),
-            inputs,
-            counts: ProofArtifactsCheckCounts {
-                selected: report.selected,
-                executed: report.executed,
-            },
-            execution_status: Some(report.execution_status.clone()),
-            guard_reason: Some(report.guard_reason.clone()),
-            errors: Vec::new(),
-        },
-        Err(error) => ProofArtifactsCheckReceipt {
-            schema: schema.to_string(),
-            ok: false,
-            verifier: verifier.to_string(),
-            inputs,
-            counts: ProofArtifactsCheckCounts {
-                selected: 0,
-                executed: 0,
-            },
-            execution_status: None,
-            guard_reason: None,
-            errors: vec![error.to_string()],
-        },
-    }
-}
-
-fn proof_run_check_receipt(
-    args: &ProofRunArtifactsCheckArgs,
-    outcome: &std::result::Result<ProofRunArtifactsReport, anyhow::Error>,
-) -> ProofArtifactsCheckReceipt {
-    let mut inputs = BTreeMap::new();
-    inputs.insert(
-        "proof_run_summary".to_string(),
-        args.proof_run_summary.to_string_lossy().to_string(),
-    );
-
-    match outcome {
-        Ok(report) => ProofArtifactsCheckReceipt {
-            schema: PROOF_RUN_ARTIFACTS_CHECK_SCHEMA.to_string(),
-            ok: true,
-            verifier: "proof-run-artifacts-check".to_string(),
-            inputs,
-            counts: ProofArtifactsCheckCounts {
-                selected: report.executed,
-                executed: report.executed,
-            },
-            execution_status: Some("executed".to_string()),
-            guard_reason: Some(report.guard_reason.clone()),
-            errors: Vec::new(),
-        },
-        Err(error) => ProofArtifactsCheckReceipt {
-            schema: PROOF_RUN_ARTIFACTS_CHECK_SCHEMA.to_string(),
-            ok: false,
-            verifier: "proof-run-artifacts-check".to_string(),
-            inputs,
-            counts: ProofArtifactsCheckCounts {
-                selected: 0,
-                executed: 0,
-            },
-            execution_status: None,
-            guard_reason: None,
-            errors: vec![error.to_string()],
-        },
-    }
-}
-
-fn write_check_receipt(path: Option<&Path>, receipt: &ProofArtifactsCheckReceipt) -> Result<()> {
-    if let Some(path) = path {
-        write_text(path, &serde_json::to_string_pretty(receipt)?)?;
-    }
-    Ok(())
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-struct ProofRunObservation {
-    schema: String,
-    status: String,
-    execution_status: String,
-    profile: String,
-    base: String,
-    head: String,
-    ok: bool,
-    execution_guard: ProofRunObservationGuard,
-    counts: ProofRunObservationCounts,
-    scopes: Vec<ProofRunObservationScope>,
-    changed_files: Vec<String>,
-    unknown_files: Vec<String>,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-struct ProofRunObservationGuard {
-    enabled: bool,
-    ci: bool,
-    reason: String,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-struct ProofRunObservationCounts {
-    commands_total: usize,
-    required_planned: usize,
-    advisory_skipped: usize,
-    executed: usize,
-    passed: usize,
-    failed: usize,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-struct ProofRunObservationScope {
-    name: String,
-    kind: String,
-    command: String,
-    status: String,
-    exit_code: Option<i64>,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-struct ProofRunObservationCollection {
-    schema: String,
-    ok: bool,
-    counts: ProofRunObservationCollectionCounts,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    window: Option<ProofRunObservationWindow>,
-    profiles: Vec<ProofRunObservationProfileSummary>,
-    scopes: Vec<ProofRunObservationScopeSummary>,
-    guards: Vec<ProofRunObservationGuardSummary>,
-    sources: Vec<ProofRunObservationSourceSummary>,
-}
-
-#[derive(Debug, Default, Serialize, PartialEq, Eq)]
-struct ProofRunObservationCollectionCounts {
-    observations: usize,
-    commands_total: usize,
-    required_planned: usize,
-    advisory_skipped: usize,
-    executed: usize,
-    passed: usize,
-    failed: usize,
-    unknown_files: usize,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-struct ProofRunObservationWindow {
-    source: String,
-    expected_runs: usize,
-    observed_runs: usize,
-    missing_runs: usize,
-    unmatched_observations: usize,
-    missing: Vec<ProofExecutorSourceRun>,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-struct ProofRunObservationProfileSummary {
-    profile: String,
-    observations: usize,
-    required_planned: usize,
-    executed: usize,
-    passed: usize,
-    failed: usize,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-struct ProofRunObservationScopeSummary {
-    name: String,
-    kind: String,
-    observations: usize,
-    executed: usize,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-struct ProofRunObservationGuardSummary {
-    reason: String,
-    observations: usize,
-    ci_observations: usize,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-struct ProofRunObservationSourceSummary {
-    path: String,
-    status: String,
-    execution_status: String,
-    profile: String,
-    base: String,
-    head: String,
-    guard_reason: String,
-    commands_total: usize,
-    required_planned: usize,
-    executed: usize,
-    passed: usize,
-    failed: usize,
-}
-
-#[derive(Debug, Clone, Copy)]
-struct ExecutionStateContext<'a> {
-    execution_status: &'a str,
-    guard_enabled: bool,
-    selected: usize,
-    executed: usize,
-    artifact_root: Option<&'a Path>,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-struct ProofExecutionObservation {
-    schema: String,
-    status: String,
-    execution_status: String,
-    profile: String,
-    base: String,
-    head: String,
-    family: String,
-    required: bool,
-    ok: bool,
-    execution_guard: ProofExecutionObservationGuard,
-    counts: ProofExecutionObservationCounts,
-    scopes: Vec<ProofExecutionObservationScope>,
-    changed_files: Vec<String>,
-    unknown_files: Vec<String>,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-struct ProofExecutionObservationGuard {
-    enabled: bool,
-    ci: bool,
-    reason: String,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-struct ProofExecutionObservationCounts {
-    selected: usize,
-    executed: usize,
-    passed: usize,
-    failed: usize,
-    artifacts: usize,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-struct ProofExecutionObservationScope {
-    name: String,
-    kind: String,
-    command: String,
-    artifact_path: Option<String>,
-    status: String,
-    exit_code: Option<i64>,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-struct ProofExecutionObservationCollection {
-    schema: String,
-    ok: bool,
-    counts: ProofExecutionObservationCollectionCounts,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    window: Option<ProofExecutionObservationWindow>,
-    families: Vec<ProofExecutionObservationFamilySummary>,
-    scopes: Vec<ProofExecutionObservationScopeSummary>,
-    sources: Vec<ProofExecutionObservationSourceSummary>,
-}
-
-#[derive(Debug, Default, Serialize, PartialEq, Eq)]
-struct ProofExecutionObservationCollectionCounts {
-    observations: usize,
-    selected: usize,
-    executed: usize,
-    passed: usize,
-    failed: usize,
-    artifacts: usize,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-struct ProofExecutionObservationWindow {
-    source: String,
-    expected_runs: usize,
-    observed_runs: usize,
-    missing_runs: usize,
-    unmatched_observations: usize,
-    missing: Vec<ProofExecutorSourceRun>,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-struct ProofExecutionObservationFamilySummary {
-    family: String,
-    observations: usize,
-    selected: usize,
-    executed: usize,
-    passed: usize,
-    artifacts: usize,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-struct ProofExecutionObservationScopeSummary {
-    name: String,
-    kind: String,
-    family: String,
-    observations: usize,
-    executed: usize,
-    artifacts: usize,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-struct ProofExecutionObservationSourceSummary {
-    path: String,
-    status: String,
-    execution_status: String,
-    profile: String,
-    base: String,
-    head: String,
-    family: String,
-    guard_reason: String,
-    selected: usize,
-    executed: usize,
-    passed: usize,
-    artifacts: usize,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-struct ProofExecutorPromotionReadiness {
-    schema: String,
-    ok: bool,
-    thresholds: ProofExecutorPromotionReadinessThresholds,
-    actuals: ProofExecutorPromotionReadinessActuals,
-    collector_runs: Vec<ProofExecutorPromotionCollectorRun>,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-struct ProofExecutorPromotionReadinessThresholds {
-    min_observations: usize,
-    min_executed: usize,
-    min_scopes: usize,
-    min_artifacts: usize,
-    min_passing_collector_runs: usize,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-struct ProofExecutorPromotionReadinessActuals {
-    observations: usize,
-    executed: usize,
-    scopes: usize,
-    artifacts: usize,
-    passing_collector_runs: usize,
-}
-
-#[derive(Debug, Deserialize)]
-struct GithubRun {
-    #[serde(rename = "databaseId")]
-    database_id: u64,
-
-    #[serde(default)]
-    event: Option<String>,
-
-    #[serde(rename = "headBranch", default)]
-    head_branch: Option<String>,
-
-    #[serde(rename = "headSha", default)]
-    head_sha: Option<String>,
-
-    #[serde(rename = "createdAt", default)]
-    created_at: Option<String>,
-
-    #[serde(default)]
-    url: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
-struct ProofExecutorSourceRun {
-    database_id: u64,
-    event: Option<String>,
-    head_branch: Option<String>,
-    head_sha: Option<String>,
-    created_at: Option<String>,
-    url: Option<String>,
-}
-
-#[derive(Debug, Serialize, PartialEq, Eq)]
-struct ProofExecutorPromotionCollectorRun {
-    database_id: u64,
-    event: Option<String>,
-    head_branch: Option<String>,
-    head_sha: Option<String>,
-    created_at: Option<String>,
-    url: Option<String>,
-}
-
-#[derive(Debug)]
-struct SourcedProofRunObservation {
-    path: PathBuf,
-    observation: ProofRunObservation,
-}
-
-#[derive(Debug)]
-struct SourcedProofExecutionObservation {
-    path: PathBuf,
-    observation: ProofExecutionObservation,
-}
-
-#[derive(Default)]
-struct ProofRunProfileAccumulator {
-    observations: usize,
-    required_planned: usize,
-    executed: usize,
-    passed: usize,
-    failed: usize,
-}
-
-#[derive(Default)]
-struct ProofRunScopeAccumulator {
-    observations: usize,
-    executed: usize,
-}
-
-#[derive(Default)]
-struct ProofRunGuardAccumulator {
-    observations: usize,
-    ci_observations: usize,
-}
-
-#[derive(Default)]
-struct FamilyAccumulator {
-    observations: usize,
-    selected: usize,
-    executed: usize,
-    passed: usize,
-    artifacts: usize,
-}
-
-#[derive(Default)]
-struct ScopeAccumulator {
-    observations: usize,
-    executed: usize,
-    artifacts: usize,
-}
-
-fn read_json(path: &Path, label: &str) -> Result<Value> {
-    let raw = fs::read_to_string(path)
-        .with_context(|| format!("failed to read {label} artifact `{}`", path.display()))?;
-    serde_json::from_str(&raw)
-        .with_context(|| format!("failed to parse {label} artifact `{}`", path.display()))
-}
-
-fn artifact_root_for(summary_path: &Path) -> PathBuf {
-    summary_path
-        .parent()
-        .filter(|parent| !parent.as_os_str().is_empty())
-        .unwrap_or_else(|| Path::new("."))
-        .to_path_buf()
-}
-
-fn write_observation(path: &Path, observation: &ProofExecutionObservation) -> Result<()> {
-    write_text(path, &serde_json::to_string_pretty(observation)?)
-}
-
-fn write_proof_run_observation(path: &Path, observation: &ProofRunObservation) -> Result<()> {
-    write_text(path, &serde_json::to_string_pretty(observation)?)
-}
-
-fn write_text(path: &Path, text: &str) -> Result<()> {
-    if let Some(parent) = path.parent()
-        && !parent.as_os_str().is_empty()
-    {
-        fs::create_dir_all(parent)
-            .with_context(|| format!("failed to create `{}`", parent.display()))?;
-    }
-    fs::write(path, text).with_context(|| format!("failed to write `{}`", path.display()))
-}
-
 fn proof_run_observation(summary: &Value) -> Result<ProofRunObservation> {
     validate_proof_run_summary(summary)?;
     let entries = expect_array(
@@ -896,137 +355,6 @@ fn proof_run_observation_collection(
         source_runs_path,
         source_runs,
     ))
-}
-
-fn render_proof_run_observation_collection_markdown(
-    collection: &ProofRunObservationCollection,
-) -> String {
-    let mut out = String::new();
-    out.push_str("# Proof Run Observation Collection\n\n");
-    out.push_str("| Metric | Count |\n");
-    out.push_str("| --- | ---: |\n");
-    push_count_row(&mut out, "Observations", collection.counts.observations);
-    push_count_row(
-        &mut out,
-        "Planned commands",
-        collection.counts.commands_total,
-    );
-    push_count_row(
-        &mut out,
-        "Required commands",
-        collection.counts.required_planned,
-    );
-    push_count_row(
-        &mut out,
-        "Advisory skipped commands",
-        collection.counts.advisory_skipped,
-    );
-    push_count_row(&mut out, "Executed commands", collection.counts.executed);
-    push_count_row(&mut out, "Passed commands", collection.counts.passed);
-    push_count_row(&mut out, "Failed commands", collection.counts.failed);
-    push_count_row(&mut out, "Unknown files", collection.counts.unknown_files);
-    push_count_row(&mut out, "Distinct scopes", collection.scopes.len());
-
-    if let Some(window) = &collection.window {
-        out.push_str("\n## Observation Window\n\n");
-        out.push_str(&format!("Source: `{}`\n\n", md_cell(&window.source)));
-        out.push_str("| Metric | Count |\n");
-        out.push_str("| --- | ---: |\n");
-        push_count_row(
-            &mut out,
-            "Expected successful proof runs",
-            window.expected_runs,
-        );
-        push_count_row(
-            &mut out,
-            "Observed runs with artifacts",
-            window.observed_runs,
-        );
-        push_count_row(&mut out, "Missing runs", window.missing_runs);
-        push_count_row(
-            &mut out,
-            "Unmatched observation artifacts",
-            window.unmatched_observations,
-        );
-
-        if !window.missing.is_empty() {
-            out.push_str("\n| Missing run | Branch | Created | URL |\n");
-            out.push_str("| ---: | --- | --- | --- |\n");
-            for run in &window.missing {
-                out.push_str(&format!(
-                    "| {} | `{}` | `{}` | {} |\n",
-                    run.database_id,
-                    md_cell(run.head_branch.as_deref().unwrap_or("")),
-                    md_cell(run.created_at.as_deref().unwrap_or("")),
-                    md_cell(run.url.as_deref().unwrap_or(""))
-                ));
-            }
-        }
-    }
-
-    if !collection.profiles.is_empty() {
-        out.push_str("\n## Profiles\n\n");
-        out.push_str("| Profile | Observations | Required | Executed | Passed | Failed |\n");
-        out.push_str("| --- | ---: | ---: | ---: | ---: | ---: |\n");
-        for profile in &collection.profiles {
-            out.push_str(&format!(
-                "| `{}` | {} | {} | {} | {} | {} |\n",
-                md_cell(&profile.profile),
-                profile.observations,
-                profile.required_planned,
-                profile.executed,
-                profile.passed,
-                profile.failed
-            ));
-        }
-    }
-
-    if !collection.scopes.is_empty() {
-        out.push_str("\n## Scopes\n\n");
-        out.push_str("| Scope | Kind | Observations | Executed |\n");
-        out.push_str("| --- | --- | ---: | ---: |\n");
-        for scope in &collection.scopes {
-            out.push_str(&format!(
-                "| `{}` | `{}` | {} | {} |\n",
-                md_cell(&scope.name),
-                md_cell(&scope.kind),
-                scope.observations,
-                scope.executed
-            ));
-        }
-    }
-
-    if !collection.guards.is_empty() {
-        out.push_str("\n## Guards\n\n");
-        out.push_str("| Reason | Observations | CI observations |\n");
-        out.push_str("| --- | ---: | ---: |\n");
-        for guard in &collection.guards {
-            out.push_str(&format!(
-                "| `{}` | {} | {} |\n",
-                md_cell(&guard.reason),
-                guard.observations,
-                guard.ci_observations
-            ));
-        }
-    }
-
-    if !collection.sources.is_empty() {
-        out.push_str("\n## Sources\n\n");
-        out.push_str("| Source | Profile | Executed | Passed | Guard |\n");
-        out.push_str("| --- | --- | ---: | ---: | --- |\n");
-        for source in &collection.sources {
-            out.push_str(&format!(
-                "| `{}` | `{}` | {} | {} | `{}` |\n",
-                md_cell(&source.path),
-                md_cell(&source.profile),
-                source.executed,
-                source.passed,
-                md_cell(&source.guard_reason)
-            ));
-        }
-    }
-
-    out
 }
 
 #[cfg(test)]
@@ -1224,162 +552,6 @@ fn validate_minimum(flag: &str, display_label: &str, actual: usize, required: us
     }
 
     Ok(())
-}
-
-fn render_observation_collection_markdown(
-    collection: &ProofExecutionObservationCollection,
-    args: &ProofExecutionObservationsSummaryArgs,
-) -> String {
-    let mut out = String::new();
-    out.push_str("# Proof Executor Observation Collection\n\n");
-    out.push_str("| Metric | Count |\n");
-    out.push_str("| --- | ---: |\n");
-    push_count_row(&mut out, "Observations", collection.counts.observations);
-    push_count_row(&mut out, "Selected commands", collection.counts.selected);
-    push_count_row(&mut out, "Executed commands", collection.counts.executed);
-    push_count_row(&mut out, "Passed commands", collection.counts.passed);
-    push_count_row(&mut out, "Failed commands", collection.counts.failed);
-    push_count_row(&mut out, "Artifacts", collection.counts.artifacts);
-    push_count_row(&mut out, "Distinct scopes", collection.scopes.len());
-
-    if let Some(window) = &collection.window {
-        out.push_str("\n## Observation Window\n\n");
-        out.push_str(&format!("Source: `{}`\n\n", md_cell(&window.source)));
-        out.push_str("| Metric | Count |\n");
-        out.push_str("| --- | ---: |\n");
-        push_count_row(
-            &mut out,
-            "Expected successful executor runs",
-            window.expected_runs,
-        );
-        push_count_row(
-            &mut out,
-            "Observed runs with artifacts",
-            window.observed_runs,
-        );
-        push_count_row(&mut out, "Missing runs", window.missing_runs);
-        push_count_row(
-            &mut out,
-            "Unmatched observation artifacts",
-            window.unmatched_observations,
-        );
-
-        if !window.missing.is_empty() {
-            out.push_str("\n| Missing run | Branch | Created | URL |\n");
-            out.push_str("| ---: | --- | --- | --- |\n");
-            for run in &window.missing {
-                out.push_str(&format!(
-                    "| {} | `{}` | `{}` | {} |\n",
-                    run.database_id,
-                    md_cell(run.head_branch.as_deref().unwrap_or("")),
-                    md_cell(run.created_at.as_deref().unwrap_or("")),
-                    md_cell(run.url.as_deref().unwrap_or(""))
-                ));
-            }
-        }
-    }
-
-    out.push_str("\n## Thresholds\n\n");
-    out.push_str("| Threshold | Required | Actual | Status |\n");
-    out.push_str("| --- | ---: | ---: | --- |\n");
-    push_threshold_row(
-        &mut out,
-        "Observations",
-        args.min_observations,
-        collection.counts.observations,
-    );
-    push_threshold_row(
-        &mut out,
-        "Executed commands",
-        args.min_executed,
-        collection.counts.executed,
-    );
-    push_threshold_row(
-        &mut out,
-        "Distinct scopes",
-        args.min_scopes,
-        collection.scopes.len(),
-    );
-    push_threshold_row(
-        &mut out,
-        "Artifacts",
-        args.min_artifacts,
-        collection.counts.artifacts,
-    );
-    if let Some(collector_runs_json) = &args.collector_runs_json {
-        let passing_collector_runs = read_collector_runs(collector_runs_json)
-            .map(|runs| runs.len())
-            .unwrap_or(0);
-        push_threshold_row(
-            &mut out,
-            "Passing collector runs",
-            args.min_passing_collector_runs,
-            passing_collector_runs,
-        );
-    }
-
-    if !collection.families.is_empty() {
-        out.push_str("\n## Families\n\n");
-        out.push_str("| Family | Observations | Executed | Artifacts |\n");
-        out.push_str("| --- | ---: | ---: | ---: |\n");
-        for family in &collection.families {
-            out.push_str(&format!(
-                "| `{}` | {} | {} | {} |\n",
-                md_cell(&family.family),
-                family.observations,
-                family.executed,
-                family.artifacts
-            ));
-        }
-    }
-
-    if !collection.scopes.is_empty() {
-        out.push_str("\n## Scopes\n\n");
-        out.push_str("| Scope | Kind | Family | Observations | Executed | Artifacts |\n");
-        out.push_str("| --- | --- | --- | ---: | ---: | ---: |\n");
-        for scope in &collection.scopes {
-            out.push_str(&format!(
-                "| `{}` | `{}` | `{}` | {} | {} | {} |\n",
-                md_cell(&scope.name),
-                md_cell(&scope.kind),
-                md_cell(&scope.family),
-                scope.observations,
-                scope.executed,
-                scope.artifacts
-            ));
-        }
-    }
-
-    if !collection.sources.is_empty() {
-        out.push_str("\n## Sources\n\n");
-        out.push_str("| Source | Family | Executed | Artifacts | Guard |\n");
-        out.push_str("| --- | --- | ---: | ---: | --- |\n");
-        for source in &collection.sources {
-            out.push_str(&format!(
-                "| `{}` | `{}` | {} | {} | `{}` |\n",
-                md_cell(&source.path),
-                md_cell(&source.family),
-                source.executed,
-                source.artifacts,
-                md_cell(&source.guard_reason)
-            ));
-        }
-    }
-
-    out
-}
-
-fn push_count_row(out: &mut String, label: &str, count: usize) {
-    out.push_str(&format!("| {label} | {count} |\n"));
-}
-
-fn push_threshold_row(out: &mut String, label: &str, required: usize, actual: usize) {
-    let status = if actual >= required { "ok" } else { "below" };
-    out.push_str(&format!("| {label} | {required} | {actual} | {status} |\n"));
-}
-
-fn md_cell(value: &str) -> String {
-    value.replace('|', "\\|")
 }
 
 fn collect_observation_paths(args: &ProofExecutionObservationsSummaryArgs) -> Result<Vec<PathBuf>> {
@@ -2673,109 +1845,6 @@ fn validate_command_entry(index: usize, entry: &Value, command: &Value) -> Resul
         }
     }
     Ok(())
-}
-
-fn expect_schema(value: &Value, expected: &str, label: &str) -> Result<()> {
-    expect_string_value(field(value, "schema", label)?, expected, "schema", label)
-}
-
-fn expect_equal(summary: &Value, manifest: &Value, path: &str) -> Result<()> {
-    let summary_value = field(summary, path, "executor summary")?;
-    let manifest_value = field(manifest, path, "executor manifest")?;
-    if summary_value != manifest_value {
-        bail!(
-            "executor artifact mismatch at `{path}`: summary {} != manifest {}",
-            render_json(summary_value),
-            render_json(manifest_value)
-        );
-    }
-    Ok(())
-}
-
-fn field<'a>(value: &'a Value, path: &str, label: &str) -> Result<&'a Value> {
-    let mut current = value;
-    for segment in path.split('.') {
-        current = current
-            .get(segment)
-            .with_context(|| format!("{label} artifact is missing `{path}`"))?;
-    }
-    Ok(current)
-}
-
-fn expect_array<'a>(value: &'a Value, path: &str, label: &str) -> Result<&'a Vec<Value>> {
-    value
-        .as_array()
-        .with_context(|| format!("{label} `{path}` must be an array"))
-}
-
-fn expect_bool(value: &Value, path: &str, label: &str) -> Result<bool> {
-    value
-        .as_bool()
-        .with_context(|| format!("{label} `{path}` must be a boolean"))
-}
-
-fn expect_bool_value(value: &Value, expected: bool, path: &str, label: &str) -> Result<()> {
-    let actual = expect_bool(value, path, label)?;
-    if actual != expected {
-        bail!("{label} `{path}` must be {expected}, got {actual}");
-    }
-    Ok(())
-}
-
-fn expect_string(value: &Value, path: &str, label: &str) -> Result<String> {
-    value
-        .as_str()
-        .map(ToOwned::to_owned)
-        .with_context(|| format!("{label} `{path}` must be a string"))
-}
-
-fn expect_optional_string(value: &Value, path: &str, label: &str) -> Result<Option<String>> {
-    if value.is_null() {
-        Ok(None)
-    } else {
-        expect_string(value, path, label).map(Some)
-    }
-}
-
-fn expect_string_value(value: &Value, expected: &str, path: &str, label: &str) -> Result<()> {
-    let actual = expect_string(value, path, label)?;
-    if actual != expected {
-        bail!("{label} `{path}` must be `{expected}`, got `{actual}`");
-    }
-    Ok(())
-}
-
-fn expect_string_array(value: &Value, path: &str, label: &str) -> Result<Vec<String>> {
-    let values = expect_array(value, path, label)?;
-    values
-        .iter()
-        .enumerate()
-        .map(|(index, value)| {
-            expect_string(value, &format!("{path}[{index}]"), label)
-                .with_context(|| format!("{label} `{path}` entry {index} must be a string"))
-        })
-        .collect()
-}
-
-fn expect_optional_i64(value: &Value, path: &str, label: &str) -> Result<Option<i64>> {
-    if value.is_null() {
-        return Ok(None);
-    }
-    value
-        .as_i64()
-        .map(Some)
-        .with_context(|| format!("{label} `{path}` must be an integer or null"))
-}
-
-fn expect_usize(value: &Value, path: &str, label: &str) -> Result<usize> {
-    let number = value
-        .as_u64()
-        .with_context(|| format!("{label} `{path}` must be a non-negative integer"))?;
-    usize::try_from(number).with_context(|| format!("{label} `{path}` is too large"))
-}
-
-fn render_json(value: &Value) -> String {
-    serde_json::to_string(value).unwrap_or_else(|_| "<unrenderable>".to_string())
 }
 
 #[cfg(test)]
